LayoutTests/fast/js/var-scope-chain-bug-87158-expected.txt

This test verifies finding local variable within an eval'ed function (https://bugs.webkit.org/show_bug.cgi?id=87158)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS globalA is 2
PASS successfullyParsed is true

TEST COMPLETE
